nft: move event to events mod
