+++ src/module/nft/bpf.rs
@@ -1,3 +1,19 @@
+use anyhow::Result;
+
+use crate::{
+    event_byte_array, event_section, event_section_factory,
+    events::{
+        bpf::{parse_single_raw_section, BpfRawSection},
+        nft::*,
+        *,
+    },
+    module::ModuleId,
+};
+
+// Please keep in sync with its bpf counterpart under
+// src/modules/nft/bpf/nft.bpf.c
+const NFT_NAME_SIZE: usize = 128;
+event_byte_array!(NftName, NFT_NAME_SIZE);
 /// Nft specific parameter offsets; keep in sync with its BPF counterpart in
 /// bpf/nft.bpf.c
 #[derive(Clone, Copy)]
@@ -47,3 +63,69 @@ pub(super) const VERD_STOLEN: u64 = 7;
 pub(super) const VERD_QUEUE: u64 = 8;
 pub(super) const VERD_REPEAT: u64 = 9;
 pub(super) const VERD_MAX: u64 = VERD_REPEAT;
+
+// Please keep in sync with its bpf counterpart under
+// src/modules/nft/bpf/nft.bpf.c
+#[repr(C, packed)]
+struct NftBpfEvent {
+    /// Table name.
+    tn: NftName,
+    /// Chain name.
+    cn: NftName,
+    /// Verdict.
+    v: i32,
+    /// Verdict chain name.
+    vcn: NftName,
+    /// Table handle
+    th: i64,
+    /// Chain handle
+    ch: i64,
+    /// Rule handle
+    rh: i64,
+    /// Verdict refers to the policy
+    p: u8,
+}
+
+#[derive(Default)]
+#[event_section_factory(NftEvent)]
+pub(crate) struct NftEventFactory {}
+
+impl RawEventSectionFactory for NftEventFactory {
+    fn from_raw(&mut self, raw_sections: Vec<BpfRawSection>) -> Result<Box<dyn EventSection>> {
+        let mut event = NftEvent::default();
+        let raw = parse_single_raw_section::<NftBpfEvent>(ModuleId::Nft, &raw_sections)?;
+
+        event.table_name = raw.tn.to_string()?;
+        event.chain_name = raw.cn.to_string()?;
+        event.table_handle = raw.th;
+        event.chain_handle = raw.ch;
+        event.policy = raw.p == 1;
+        event.rule_handle = match raw.rh {
+            -1 => None,
+            _ => Some(raw.rh),
+        };
+        event.verdict = match raw.v {
+            -1 => "continue",
+            -2 => "break",
+            -3 => "jump",
+            -4 => "goto",
+            -5 => "return",
+            0 => "drop",
+            1 => "accept",
+            2 => "stolen",
+            3 => "queue",
+            4 => "repeat",
+            /* NF_STOP is deprecated. */
+            5 => "stop",
+            _ => "unknown",
+        }
+        .to_owned();
+
+        // Destination chain is only valid for NFT_JUMP/NFT_GOTO.
+        if raw.v == -3 || raw.v == -4 {
+            event.verdict_chain_name = raw.vcn.to_string_opt()?;
+        }
+
+        Ok(Box::new(event))
+    }
+}
